+++ dejafu-tests/Examples/Philosophers.hs
@@ -4,13 +4,11 @@ module Examples.Philosophers where
 
 import Control.Monad (replicateM, forever)
 import Control.Monad.Conc.Class
-import Data.Functor (void)
 import Test.DejaFu
 import Test.Framework (Test)
 import Test.Framework.Providers.HUnit (hUnitTestToTests)
 import Test.HUnit (test)
 import Test.HUnit.DejaFu
-import System.Random (StdGen)
 
 tests :: [Test]
 tests = hUnitTestToTests $ test