+++ tasty-dejafu/CHANGELOG.markdown
@@ -16,7 +16,9 @@ unreleased
   function, and re-exported values.
 - Due to changes in dejafu, the `Way` type is now abstract and exposes smart constructor functions:
     - `systematically`, corresponding to the old `Systematically`.
-    - `randomly`, corresponding to the old `Randomly` with an additional parameter.
+    - `randomly`, corresponding to the old `Randomly`.
+    - `swarmy`, corresponding to the old `Randomly` and specifying how many executions to use the
+      same weights for.
 - The `defaultWay`, `defaultMemType`, and `defaultBounds` values are all now re-exported.
 
 ### Miscellaneous